Updated sentence term days to numberic with scale of 10 and precision of 2 to support values provided in Disposition extracts.  Also updated H2 and MySQL data models.

+++ adapters/analytical-data-store-persistence-adapter/src/main/java/org/ojbc/adapters/analyticaldatastore/dao/AnalyticalDatastoreDAOImpl.java
@@ -458,7 +458,7 @@ public class AnalyticalDatastoreDAOImpl implements AnalyticalDatastoreDAO{
         	            
         	            if (disposition.getSentenceTermDays() != null)
         	            {	
-        	            	ps.setInt(6, disposition.getSentenceTermDays());
+        	            	ps.setBigDecimal(6, disposition.getSentenceTermDays());
         	            }
         	            else
         	            {
